+++ src/i18next.functions.js
@@ -26,6 +26,24 @@ function addResourceBundle(lng, ns, resources, deep) {
     }
 }
 
+function hasResourceBundle(lng, ns) {
+    if (typeof ns !== 'string') {
+        ns = o.ns.defaultNs;
+    }
+
+    resStore[lng] = resStore[lng] || {};
+    var res = resStore[lng][ns] || {};
+
+    var hasValues = false;
+    for(var prop in res) {
+        if (res.hasOwnProperty(prop)) {
+            hasValues = true;
+        }
+    }
+
+    return hasValues;
+}
+
 function removeResourceBundle(lng, ns) {
     if (typeof ns !== 'string') {
         ns = o.ns.defaultNs;