+++ releasenotes.md
@@ -4,6 +4,8 @@
 - handle error on json parse when using internal xhr
 - fixes plural/singular on count if going on fallbacks eg. fr --> en
 - fixes global leak of sync in amd versions
+- apply options.lowerCaseLng to fallbackLng too
+- added hasResourceBundle(lng, ns) to check if bundle exists
 
 ### 1.7.4
 - add resource bundle gets deep extend flag i18n.addResourceBundle(lng, ns, { 'deep': { 'key2': 'value2' }}, true);